+++ app/static/lib/main.js
@@ -1,6 +1,6 @@
 // mei-friend version and date
 export const version = '1.2.4';
-export const versionDate = '3 April 2025'; // use full or 3-character english months, will be translated
+export const versionDate = '11 April 2025'; // use full or 3-character english months, will be translated
 export const splashDate = '17 January 2025'; // date of the splash screen content, same translation rules apply
 
 var vrvWorker;
@@ -96,7 +96,13 @@ import Viewer from './viewer.js';
 import * as speed from './speed.js';
 import { loginAndFetch, solid } from './solid.js';
 import Storage from './storage.js';
-import { fillInBranchContents, logoutFromGithub, refreshGithubMenu, setCommitUIEnabledStatus } from './github-menu.js';
+import {
+  fillInBranchContents,
+  logoutFromGithub,
+  refreshGithubMenu,
+  setCommitUIEnabledStatus,
+  onRemoteUpdate,
+} from './github-menu.js';
 import { forkAndOpen, forkRepositoryCancel } from './fork-repository.js';
 import {
   addZoneDrawer,
@@ -232,6 +238,18 @@ export function setMeiFileInfo(fName, fLocation, fLocationPrintable) {
 
 export function setFileLocationType(t) {
   fileLocationType = t; // wrap in function to facilitate external setting
+  if (gm) {
+    if (t === 'github') {
+      // poll for remote updates when working from github
+      gm.pollForRemoteUpdates();
+    } else {
+      // stop polling for remote updates when not working from github
+      gm.stopPollingForRemoteUpdates();
+      // remove remote change indicator
+      const remoteFileChanged = document.querySelector('#remoteFileChanged');
+      remoteFileChanged.innerHTML = '';
+    }
+  }
 }
 
 export function updateFileStatusDisplay() {
@@ -319,6 +337,7 @@ export async function updateGithubInLocalStorage() {
       githubRepo: gm.repo,
       githubToken: gm.token,
       branch: gm.branch,
+      headSha: gm.headSha,
       filepath: gm.filepath,
       userLogin: author.username,
       userName: author.name,
@@ -583,8 +602,9 @@ async function completeInitialLoad() {
         repo: storage.github.githubRepo,
         branch: storage.github.branch,
         filepath: storage.github.filepath,
+        headSha: storage.github.headSha,
+        onRemoteUpdate,
       });
-
       //document.querySelector("#fileLocation").innerText = meiFileLocationPrintable;
     } else if (storage.github && !isLoggedIn) {
       // we have github data but are not logged in
@@ -597,7 +617,7 @@ async function completeInitialLoad() {
       // let msg = await gitProxy.registerGitManager('github', 'github', githubToken);
       // console.log('Registered git manager: ', msg);
 
-      gm = new GitManager('github', 'github', githubToken);
+      gm = new GitManager('github', 'github', githubToken, { onRemoteUpdate });
       gm.getAuthor()
         .then((author) => {
           //github = new Github('', githubToken, '', '', '', userLogin, userName, userEmail);
@@ -605,6 +625,7 @@ async function completeInitialLoad() {
             githubRepo: gm.repo,
             githubToken: gm.token,
             branch: gm.branch,
+            headSha: gm.headSha,
             filepath: gm.filepath,
             userLogin: author.username,
             userName: author.name,
@@ -721,7 +742,7 @@ async function completeInitialLoad() {
     // no local storage
     if (isLoggedIn) {
       // initialise new github object
-      gm = new GitManager('github', 'github', githubToken);
+      gm = new GitManager('github', 'github', githubToken, { onRemoteUpdate });
       //github = new Github('', githubToken, '', '', '', userLogin, userName, userEmail);
     }
     meiFileLocation = '';
@@ -830,6 +851,16 @@ async function completeInitialLoad() {
       loginAndFetch(getSolidIdP(), populateSolidTab);
     }, restoreSolidTimeoutDelay);
   }
+
+  // if we are working from github and have a gm, clone it
+  if (fileLocationType === 'github' && gm) {
+    // remove remote changed indicator
+    const remoteFileChanged = document.getElementById('remoteFileChanged');
+    if (remoteFileChanged) {
+      remoteFileChanged.innerHTML = '';
+    }
+    gm.clone();
+  }
 } // completeInitialLoad()
 
 export async function openUrlFetch(url = '', updateAfterLoading = true) {
@@ -848,9 +879,15 @@ export async function openUrlFetch(url = '', updateAfterLoading = true) {
 
       if (fileLocationType === 'github' && userOrg && repo && branch && filepath) {
         // clone repo
-        gm = new GitManager('github', 'github', githubToken);
+        gm = new GitManager('github', 'github', githubToken, { onRemoteUpdate });
         // TODO modify for multiple git providers
         // TODO use checkAndClone mechanism to warn about excessive sizes
+
+        // remove remote changed indicator
+        const remoteFileChanged = document.getElementById('remoteFileChanged');
+        if (remoteFileChanged) {
+          remoteFileChanged.innerHTML = '';
+        }
         gm.clone(`https://github.com/${userOrg}/${repo}.git`, branch)
           .then(() => {
             gm.readFile(filepath)
@@ -919,6 +956,7 @@ function openUrlProcess(content, url, updateAfterLoading) {
     // re-initialise github menu since we're now working from a URL
     gm.filepath = '';
     gm.branch = '';
+    gm.headSha = '';
     if (storage.supported) {
       updateGithubInLocalStorage();
     }
@@ -1215,7 +1253,7 @@ export function openFile(file = defaultMeiFileURL, setFreshlyLoaded = true, upda
   if (storage.supported) {
     storage.fileLocationType = 'url';
   }
-  fileLocationType = 'url';
+  setFileLocationType('url');
   if (gm) gm.filepath = '';
   if (file === defaultMeiFileURL) {
     // opening default MEI file