+++ CHANGELOG.md
@@ -5,7 +5,7 @@
   * Known shortcomings: 
       - missing support for markup elements `choice`, `subst`
       - `meterSigGrp` not handled (`func="mixed,alternating"`)
-
+* Improve handling of remote changes to Git-managed files
 
 ### 1.2.3 patch
 * Add latest Vervio release 5.1.0